<Dav999> Oh yeah, something about integers

+++ main.py
@@ -1545,8 +1545,8 @@ async def on_message(message):
 		# what kind of stupid language uses elif instead of elseif or else if?
 			try:
 				cmdbits = arguments.split() # should split it so [1] is number, [2] is operand, [3] is third number
-				cmdbits[0] = int(cmdbits[0])
-				cmdbits[2] = int(cmdbits[2])
+				cmdbits[0] = float(cmdbits[0])
+				cmdbits[2] = float(cmdbits[2])
 				# But apparently not, it's [0] / [1] / [2] instead
 				out = "" # setting extra crashes
 				if cmdbits[1] == "+":
